Changed recaudacion and objetivorecaudacion type to float (#94)

+++ src/main/java/Datos/Evento/EventoConcierto.java
@@ -8,17 +8,17 @@ public class EventoConcierto extends Evento {
     // Empty builder, should not be used. Used for json deserialization purposes exclusively by Jackson
     public EventoConcierto(){}
 
-    public EventoConcierto(String nombre, String ubicacion, int objetivoRecaudacion, String descripcion, String date, String url, String artista) {
+    public EventoConcierto(String nombre, String ubicacion, float objetivoRecaudacion, String descripcion, String date, String url, String artista) {
         super(nombre, ubicacion, objetivoRecaudacion, descripcion, date, url);
         this.artista = artista;
     }
 
-    public EventoConcierto(String nombre, String ubicacion, int recaudacion, int objetivoRecaudacion, String descripcion, String date, String url, String artista) {
+    public EventoConcierto(String nombre, String ubicacion, float recaudacion, float objetivoRecaudacion, String descripcion, String date, String url, String artista) {
         super(nombre, ubicacion, recaudacion, objetivoRecaudacion, descripcion, date, url);
         this.artista = artista;
     }
 
-    public EventoConcierto(String nombre, String ubicacion, int recaudacion, int objetivoRecaudacion, String descripcion, String date, String url, int ID, String artista) {
+    public EventoConcierto(String nombre, String ubicacion, float recaudacion, float objetivoRecaudacion, String descripcion, String date, String url, int ID, String artista) {
         super(nombre, ubicacion, recaudacion, objetivoRecaudacion, descripcion, date, url, ID);
         this.artista = artista;
     }
